\name{emon3}
\alias{emon.3}
\alias{emon3}
\docType{data}
\title{EMON Dataset}
\description{
    The Lake Pomona diseaster response network that is used for
    illustrating the graphletCov terms. The dataset is originally
    known as EMON dataset in literature. However, we modified the
    dataset to make it undirected and included some categorical
    node attributes that were originally provided with the network.
}
\usage{emon3}
\format{A graph object containing a set of node attributes}
\source{Drabek et al. (1981), Managing Multiorganizational Emergency Responses: Emergent Search 
and Rescue Networks in Natural Disaster and Remote Area Settings}
\references{
Yaveroglu ON, Fitzhugh SM, Kurant M, Markopoulou A, Butts CT, Przulj N (2013).
{\pkg{ergm.graphlets}: {A} Package for ERG Modeling Based on Graphlet Properties}, 
Journal of Statistical Software 1(1), 1-2,
URL http://www.jstatsoft.org/.
}
\seealso{statnet, network, ergm, ergm-terms}
\keyword{datasets}
